Stage, LinearStage, CSSStage, LinearCSSStage
Stage and siblings build on TransitionGroup to help build flows of
content that can be stepped through.

One use is to build an array of React elements that take over an
application. Every time an item is added or removed from that array
create a wrapping Stage around it and render that stage in a parent
element. When the visible item in the stage is complete it can call a
passed shouldLeave function or something in the containing parent to
have the array modified so the next item is displayed. This can be
great for controlling tutorial flows.

LinearStage takes away some of the control while doing some of this for
you. In cases where the flow is less dynamically created a LinearStage
can wrapped a series of known steps and render them in order allowing
it to be used to produce key frame like steps of high level react and
combined with Animated to animate between those key frames.