Invoke Karma directly and add CI-mode npm script
Using grunt-karma as an intermediary does not gain us anything in
terms of performance, and actually complicates the configuration
because several parameters must be explicitly set in both the karma
config _and_ grunt config in order for the tests to run as expected.

This PR removes grunt-karma and aliases `npm run test-dev` to execute
Karma directly.

It also aliases `npm test` to Karma in --single-run mode, which runs
the tests in Firefox (PhantomJS is unsuitable due to the dependence
on modern DOM features) and then exits ('continuous integration' mode)