Silence uncaught errors in timers given to AnimatedAgent
Errors are still passed normally in Timers so that apps can listen for
them, the normal uncaught error error will just be silent in. Probably
will revisit and consider a way to exclude this behaviour in dev but
the likely case is that the uncaught error error would occur very
frequently and restrict use of dev tools that report it.