Move style storage and logic into Animated
AnimatedAgent is doing too much. Move the storage of replaced style
into Animated. In Animated that data will be cleaned up with the
element instance when its unmounted instead of needing to worry about
adding special logic to Agent. This also provides a more clear cut of
responsibilities for handling more of the rendering changes an
animation makes like adding methods that handle attributes.

- Add Animated.restoreStyle, performs the replaceStyle with no
  arguments logic providing a clearer name.