Untrack id and member expressions that are assigned to
Replace nodes with the id or member expression they were last set to
until such id or member expression is set. In other words notice when a
value is set and use the previous assignment, since its a saved
reference to the last value. This will dead code reduce many
unnecessary assignments.