Update preact crawler tests with destroy delay
Crawler waits a javascript frame before destroying an animated state.
This way it can be any time in that javascript frame that another
element will claim the animated state. This is important in case the
original element and new element happen in the same update but the new
element comes after the old element was destroyed.